extracted format date to date.ts helper
