+++ src/helpers/date.ts
@@ -25,3 +25,45 @@ export const formatDate = ({
     }),
   }).format(dateObj);
 };
+
+/**
+ * Formats transaction date for display in transaction contexts
+ * Can be used with or without time component
+ * @param createdAt - ISO date string or undefined for current time
+ * @param includeTime - Whether to include time (default: true)
+ * @returns Formatted date string
+ */
+export const formatTransactionDate = (
+  createdAt?: string,
+  includeTime: boolean = true,
+): string => {
+  let dateObj: Date;
+
+  if (createdAt) {
+    dateObj = new Date(createdAt);
+  } else {
+    dateObj = new Date();
+  }
+
+  if (!includeTime) {
+    // Simple format for history lists: "Dec 13"
+    return dateObj.toDateString().split(" ").slice(1, 3).join(" ");
+  }
+
+  // Comprehensive format for transaction details: "Dec 13, 2023 · 2:30pm"
+  const formattedDate = dateObj.toLocaleDateString("en-US", {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  });
+
+  const formattedTime = dateObj
+    .toLocaleTimeString("en-US", {
+      hour: "numeric",
+      minute: "2-digit",
+      hour12: true,
+    })
+    .toLowerCase();
+
+  return `${formattedDate} · ${formattedTime}`;
+};
